Add utils and test for it
